refactor: inline hooks creators
